feat: adicionada a requisição para buscar o retrospecto das partidas de um determinado time

+++ src/main/java/com/waynesouza/soccer/domain/dto/PartidaDTO.java
@@ -7,7 +7,6 @@ import lombok.Getter;
 import lombok.Setter;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -22,7 +21,7 @@ public class PartidaDTO implements Serializable {
     @NotNull(message = "A quantidade de gols do time mandante deve ser informada")
     @Min(value = 0, message = "O valor deve ser um número inteiro maior que zero")
     @Digits(integer = 2, fraction = 0, message = "O valor deve ser um número inteiro positivo ou zero")
-    private BigDecimal quantidadeGolMandante;
+    private Integer quantidadeGolMandante;
 
     @NotNull(message = "O time visitante deve ser informado")
     private String timeVisitante;
